backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
ultralytics==8.3.0
Pillow==11.1.0
numpy==2.2.2
gunicorn==23.0.0
python-dotenv==1.0.1
